COLLISION OPTIMIZATION (must read below)
if problems with collisions you may revert to commits ago, this commit has :
cleaning and organizing of map input of polylines (Addpolyline function)
deleting iterations through all colliders to all colliders, leaving only the player collider that checks if he has collision with something. So collisions different to PLAYER to SOMETHING (SOMETHING to SOMETHING) may not work from this commit.